requirements.txt
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.1
pandas==2.1.3
plotly==5.17.0
numpy==1.24.3
